Align Hall of Fame lines correctly
Now that the rank (1..10) width varies on the lines in the Hall of Fame,
put the rank in a fixed-width item. One-digit ranks are prefixed with a
space to align the rank more in the middle instead of the left.

Also updated the initial score values to my current scores.